+++ packages/fireproof/test/www/todo.html
@@ -5,7 +5,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Fireproof Test</title>
-  <script src="../../dist/browser/fireproof.iife.js?64"></script>
+  <script src="../../dist/browser/fireproof.iife.js?69"></script>
   <script type="text/javascript">
     function todoApp() {
       const actorTag = Math.random().toString(36).substring(2, 7)
@@ -27,7 +27,7 @@
         // meta in w3, storage in web3
         // connection = connect.s3(db, s3Info)
 
-        connection = db.connect("jchris+todos-2@fireproof.storage", "todo-test")
+        connection = db.connect("jchris+todos-5@fireproof.storage", "todo-test")
         window.connection = connection
         cancelRefresh = autoRefresh()
